+++ src/core/serialize/interactions.js
@@ -14,6 +14,7 @@ written permission of Adobe.
 const xd = require("scenegraph");
 
 const $ = require("../../utils/utils");
+const { trace } = require("../../utils/trace");
 const PropType = require("../proptype");
 const NodeUtils = require("../../utils/nodeutils");
 
@@ -27,98 +28,82 @@ function getSizedGestureDetector(xdNode, serializer, ctx, paramName, isOwn) {
 }
 exports.getSizedGestureDetector = getSizedGestureDetector;
 
-function getPageLink(xdNode, serializer, ctx, child) {
-	if (NodeUtils.getProp(xd.root, PropType.ENABLE_PROTOTYPE) && xdNode.triggeredInteractions.length) {
-		let interaction = xdNode.triggeredInteractions[0];
-		if (xdNode.triggeredInteractions.length > 1) {
-			ctx.log.warn(`Multiple interactions on one object is not unsupported.`, xdNode);
-		}
-		if (interaction.trigger.type !== "tap") {
-			ctx.log.warn(`Trigger unsupported ('${interaction.trigger.type}'), only tap triggered are supported.`, xdNode);
-		}
-		let builder = "";
-		let duration = "";
-		let ease = "";
-		let transition = "";
-		if (interaction.action.type === "goToArtboard") {
-			if (interaction.action.transition.type === "slide") {
-				switch (interaction.action.transition.fromSide) {
-					case "L":
-						transition = "transition: LinkTransition.SlideLeft";
-						break;
-					case "R":
-						transition = "transition: LinkTransition.SlideRight";
-						break;
-					case "T":
-						transition = "transition: LinkTransition.SlideDown";
-						break;
-					case "B":
-						transition = "transition: LinkTransition.SlideUp";
-						break;
-				}
-			} else if (interaction.action.transition.type === "push") {
-				switch (interaction.action.transition.fromSide) {
-					case "L":
-						transition = "transition: LinkTransition.PushLeft";
-						break;
-					case "R":
-						transition = "transition: LinkTransition.PushRight";
-						break;
-					case "T":
-						transition = "transition: LinkTransition.PushDown";
-						break;
-					case "B":
-						transition = "transition: LinkTransition.PushUp";
-						break;
-				}
-			} else if (interaction.action.transition.type === "dissolve") {
-				transition = "transition: LinkTransition.Fade";
-			} else {
-				ctx.log.warn(`Transition type not supported: ${interaction.action.transition.type}.`, xdNode);
-			}
-			let artboard = ctx.getArtboardFromXdNode(interaction.action.destination);
-			builder = `pageBuilder: () => ${artboard.toString(serializer, ctx)}`;
-			duration = `duration: ${$.fix(interaction.action.transition.duration, 4)}`;
-			switch (interaction.action.transition.easing) {
-				case "linear":
-					ease = "ease: Curves.linear";
-					break;
-				case "ease-in":
-					ease = "ease: Curves.easeIn";
-					break;
-				case "ease-out":
-					ease = "ease: Curves.easeOut";
-					break;
-				case "ease-in-out":
-					ease = "ease: Curves.easeInOut";
-					break;
-				case "wind-up":
-					ease = "ease: Curves.slowMiddle"; // TODO: CE: Find correct ease function
-					break;
-				case "bounce":
-					ease = "ease: Curves.bounceIn";
-					break;
-				case "snap":
-					ease = "ease: Curves.easeInOutExpo"; // TODO: CE: Find correct ease function
-					break;
-				default:
-					ctx.log.warn(`Ease '${interaction.action.transition.easing}' unsupported.`, xdNode);
-					ease = "";
-					break;
-			}
-		} else if (interaction.action.type === "goBack") {
-			builder = "";
-		} else {
-			ctx.log.warn(`Action type '${interaction.action.type}' unsupported.`, xdNode);
-		}
-		if (transition) transition += ",";
-		if (duration) duration += ",";
-		if (ease) ease += ",";
-		if (builder) builder += ",";
-		let str = `PageLink(links: [PageLinkInfo(${transition}${duration}${ease}${builder}), ], child: ${child}, )`;
-		return str;
-	} else {
-		return child;
+function getPageLink(xdNode, serializer, ctx, childStr) {
+	if (!NodeUtils.getProp(xd.root, PropType.ENABLE_PROTOTYPE) || !xdNode.triggeredInteractions.length) {
+		return childStr;
 	}
+
+	let interaction = xdNode.triggeredInteractions[0];
+	if (xdNode.triggeredInteractions.length > 1) {
+		ctx.log.warn(`Multiple interactions on one object is not unsupported.`, xdNode);
+		return childStr;
+	}
+	if (interaction.trigger.type !== "tap") {
+		ctx.log.warn(`Trigger unsupported ('${interaction.trigger.type}'), only tap triggered are supported.`, xdNode);
+		return childStr;
+	}
+
+	let action = interaction.action, transition = action.transition;
+	let builderParam = "", durationParam = "", easeParam = "", transitionParam = "";
+	if (action.type === "goToArtboard") {
+		let artboard = ctx.getArtboardFromXdNode(action.destination);
+		if (!artboard) {
+			trace(`Didn't add PageLink for destination artboard '${action.destination.name}' because it was not found. This is likely due to a duplicate name.`);
+			return childStr;
+		}
+
+		transitionParam = _getTransitionParam(transition);
+		if (!transitionParam) { ctx.log.warn(`Transition type not supported: '${transition.type}'.`, xdNode); }
+		easeParam = _getEaseParam(transition);
+		if (!easeParam) { ctx.log.warn(`Ease not supported: '${transition.easing}'.`, xdNode); }
+
+		builderParam = `pageBuilder: () => ${artboard.toString(serializer, ctx)}, `;
+		durationParam = _getDurationParam(transition);
+	} else if (action.type === "goBack") {
+		// do nothing, PageLink treats an empty builder as "go back".
+	} else {
+		ctx.log.warn(`Action type '${action.type}' unsupported.`, xdNode);
+		return childStr;
+	}
+	return `PageLink(links: [PageLinkInfo(${transitionParam}${easeParam}${durationParam}${builderParam}), ], child: ${childStr}, )`;
 }
 exports.getPageLink = getPageLink;
+
+
+function _getTransitionParam(transition) {
+	let type = TRANSITION_TYPE_MAP[transition.type] || "";
+	let dir = TRANSITION_SIDE_MAP[transition.fromSide] || "";
+	return type && `transition: LinkTransition.${type}${dir}, `;
+}
+
+function _getEaseParam(transition) {
+	let ease = TRANSITION_EASE_MAP[transition.easing] || "";
+	return ease && `ease: Curves.${ease}, `;
+}
+
+function _getDurationParam(transition) {
+	return `duration: ${$.fix(transition.duration, 4)}, `;
+}
+
+const TRANSITION_TYPE_MAP = {
+	"slide": "Slide",
+	"push": "Push",
+	"dissolve": "Fade",
+}
+
+const TRANSITION_SIDE_MAP = {
+	"L": "Left",
+	"R": "Right",
+	"T": "Up",
+	"B": "Down"
+}
+
+const TRANSITION_EASE_MAP = {
+	"linear": "linear",
+	"ease-in": "easeIn",
+	"ease-out": "easeOut",
+	"ease-in-out": "easeInOut",
+	"wind-up": "slowMiddle",
+	"bounce": "bounceIn",
+	"snap": "easeInOutExpo",
+}